chore: checkpoint token scale refresh
Includes 2026-05-31 token telemetry, non-specialist scale analogies, and current parallel-agent workspace changes. Excludes protected 1334 paths and root CON.

+++ Assets/_Project/Scripts/World/FloraAmbientSway/FloraAmbientSwayRuntime.cs
@@ -1377,7 +1377,7 @@ namespace Hecton8.World.FloraAmbientSway
         {
             buffer = default;
             return vault != null &&
-                   handle.BufferID == expectedBufferId &&
+                   handle.BufferID == (uint)expectedBufferId &&
                    !vault.IsCompactionFenceActive &&
                    vault.TryResolveHandle(in handle, out buffer) &&
                    buffer.IsCreated &&
